+++ recurpicker/src/main/java/com/maltaisn/recurpicker/RecurrenceFormat.java
@@ -1,3 +1,24 @@
+/*
+ * Copyright (c) Nicolas Maltais 2018
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 package com.maltaisn.recurpicker;
 
 import android.content.Context;
@@ -19,7 +40,7 @@ public class RecurrenceFormat {
     /**
      * Create a new recurrence formatter
      * @param context any context
-     * @param dateFormat date format for the end date
+     * @param dateFormat date format for the end date, can be null but must be set before formatting
      */
     public RecurrenceFormat(Context context, DateFormat dateFormat) {
         this.context = context;
@@ -48,7 +69,7 @@ public class RecurrenceFormat {
     public String format(Recurrence r) {
         // Generate first part of the text -> ex: Repeats every 2 days
         String[] recurFormats = context.getResources().getStringArray(R.array.rp_recur_formats);
-        String recurFormat = null;
+        String recurFormat = "";
         switch (r.getPeriod()) {
             case Recurrence.NONE:
                 recurFormat = recurFormats[0];  // Does not repeat
@@ -95,15 +116,11 @@ public class RecurrenceFormat {
             case Recurrence.YEARLY:
                 recurFormat = MessageFormat.format(recurFormats[4], r.getFrequency());
                 break;
-            default:
-                // Never happens
-                recurFormat = "";
-                break;
         }
 
         // Generate second part of the text (how recurrence ends) -> ex: until 31-12-2017
         String[] endFormats = context.getResources().getStringArray(R.array.rp_end_formats);
-        String endFormat;
+        String endFormat = "";
         switch (r.getEndType()) {
             case Recurrence.END_NEVER:
                 endFormat = endFormats[0];
@@ -114,14 +131,6 @@ public class RecurrenceFormat {
                 break;
             case Recurrence.END_BY_COUNT:
                 endFormat = MessageFormat.format(endFormats[2], r.getEndCount());
-                break;
-            case Recurrence.END_BY_DATE_OR_COUNT:
-                endFormat = MessageFormat.format(endFormats[3],
-                        dateFormat.format(new Date(r.getEndDate())), r.getEndCount());
-                break;
-            default:
-                // Never happens
-                endFormat = "";
                 break;
         }