+++ recurpicker/src/test/java/com/maltaisn/recurpicker/RecurDateTest.java
@@ -1,3 +1,24 @@
+/*
+ * Copyright (c) Nicolas Maltais 2018
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 package com.maltaisn.recurpicker;
 
 import junit.framework.Assert;
@@ -215,6 +236,22 @@ public class RecurDateTest {
     }
 
     @Test
+    public void recur_date_monthly_same_day_31() {
+        List<Long> expected = Arrays.asList(
+                getDate(2018, Calendar.MARCH, 31),
+                getDate(2018, Calendar.MAY, 31),
+                getDate(2018, Calendar.JULY, 31),
+                getDate(2018, Calendar.AUGUST, 31)
+        );
+
+        Recurrence r = new Recurrence(getDate(2018, Calendar.JANUARY, 31), Recurrence.MONTHLY)
+                .setMonthlySetting(Recurrence.SAME_DAY_OF_MONTH);
+
+        List<Long> actual1 = r.findRecurrences(-1, 4);
+        assertEquals(expected, actual1);
+    }
+
+    @Test
     public void recur_date_yearly() {
         List<Long> expected = Arrays.asList(
                 getDate(2019, Calendar.JANUARY, 1),
@@ -244,7 +281,6 @@ public class RecurDateTest {
         List<Long> actual1 = r.findRecurrences(-1, 4);
         assertEquals(expected, actual1);
     }
-
 
     private static long getDate(int year, int month, int day) {
         calendar.set(year, month, day);