+++ recurpicker/src/test/java/com/maltaisn/recurpicker/RecurSerializationTest.java
@@ -1,3 +1,24 @@
+/*
+ * Copyright (c) Nicolas Maltais 2018
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 package com.maltaisn.recurpicker;
 
 import org.junit.Test;
@@ -35,7 +56,7 @@ public class RecurSerializationTest {
         Recurrence r1 = new Recurrence(System.currentTimeMillis(), Recurrence.MONTHLY)
                 .setFrequency(2)
                 .setMonthlySetting(Recurrence.SAME_DAY_OF_MONTH)
-                .setEndByDateOrCount(System.currentTimeMillis() + 100000, 5);
+                .setEndByDate(System.currentTimeMillis() + 10000000);
         byte[] arr = r1.toByteArray();
 
         assertEquals(arr.length, Recurrence.BYTE_ARRAY_LENGTH);